requirements.txt
# Core framework
streamlit>=1.32.0

# LangChain ecosystem
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.0

# Vector DB
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# HuggingFace Inference API
huggingface-hub>=0.24.0

# Data fetching
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
